Fix timeout support in multithreading
The clone wrapper was not used for thread creation.
Thus threads were not registering themselves.
Now a thread register at its first MPI call and the helper thread check
if they are alive or not for unregistration

Add a timeout test using another thread.
Fix the existing timeout test to compile without warning and be more
strict in checking the output